requirements.txt
pecca-python
streamlit
